pyproject.toml
[build-system]
requires = ["setuptools>=80", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "pyturbocode"
dynamic = ["version"]
description = "A simple implementation of a turbo encoder and decoder"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.2",
    "scikit-commpy>=0.8.0",
]
authors = [
  {name = "Yann de Thé", email = "ydethe@gmail.com"},
]

[project.urls]
"Bug Tracker" = "https://github.com/ydethe/turbocode/issues"
Homepage = "https://github.com/ydethe/turbocode"
Source = "https://github.com/ydethe/turbocode"

[tool.setuptools_scm]
# version_scheme = "no-guess-dev"
local_scheme = "no-local-version"

[dependency-groups]
dev = [
    "coverage-badge>=1.1.0",
    "ipython>=8.30.0",
    "pre-commit>=4.0.1",
    "black>=22.3.0",
]
doc = [
    "pdoc3>=0.11.3",
    "genbadge[all]>=1.1.1,<2.0.0",
    "docstr-coverage>=2.2.0",
]
test = [
    "pytest-asyncio>=0.25.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-html>=4.1.1",
    "pytest-instafail>=0.5.0",
    "pytest-mock>=3.14.0",
    "pytest-picked>=0.5.1",
    "pytest-sugar>=1.0.0",
    "pytest-xdist>=3.6.1",
]

[tool.pylint.format]
max-line-length = 100

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = [
    "--import-mode=importlib",
    "--html=htmldoc/tests/report.html",
    "--self-contained-html",
    "--junitxml=htmldoc/junit.xml",
    "--cov-config=tests/coverage.conf",
    "--cov=pyturbocode",
    "--doctest-modules",

]
testpaths = [
    "tests/",
]

[tool.pdm.scripts]
doc.shell = """\
    mkdir -p htmldoc/pyturbocode
    pdoc --html --force --config latex_math=True -o htmldoc pyturbocode
    coverage html -d htmldoc/coverage --rcfile tests/coverage.conf
    coverage xml -o htmldoc/coverage/coverage.xml --rcfile tests/coverage.conf
    docstr-coverage src/pyturbocode -miP -sp -is -idel --skip-file-doc --badge=htmldoc/pyturbocode/doc_badge.svg
    genbadge coverage -l -i htmldoc/coverage/coverage.xml -o htmldoc/pyturbocode/cov_badge.svg
"""
